Can evalue user rollout and toggle a flag correctly based on hashed user context string

+++ lib/types/types.go
@@ -6,16 +6,16 @@ type Flag struct {
 	Title string `json:"title"`
 	Description string `json:"description"`
 	IsActive bool `json:"is_active"`
-	RolloutPercentage int `json:"rollout_percentage"`
+	RolloutPercentage float32 `json:"rollout_percentage"`
 	WhiteListedUsers string `json:"white_listed_users"`
-	ErrorThresholdPercentage int `json:"error_threshold_percentage"`
+	ErrorThresholdPercentage float32 `json:"error_threshold_percentage"`
 	CircuitStatus string `json:"circuit_status"`
 	IsRecoverable bool `json:"is_recoverable"`
-	CircuitRecoveryPercentage int `json:"circuit_recovery_percentage"`
+	CircuitRecoveryPercentage float32 `json:"circuit_recovery_percentage"`
 	CircuitRecoveryDelay int `json:"circuit_recovery_delay"`
-	CircuitInitialRecoveryPercentage int `json:"circuit_initial_recovery_percentage"`
+	CircuitInitialRecoveryPercentage float32 `json:"circuit_initial_recovery_percentage"`
 	CircuitRecoveryRate int `json:"circuit_recovery_rate"`
-	CircuitRecoveryIncrementPercentage int `json:"circuit_recovery_increment_percentage"`
+	CircuitRecoveryIncrementPercentage float32 `json:"circuit_recovery_increment_percentage"`
 	CircuitRecoveryProfile string `json:"circuit_recovery_profile"`
 	CreatedAt string `json:"created_at"`
 	UpdatedAt string `json:"updated_at"`
